Removes all references to swf files (used by old flash client)
