+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/presentation/PreuploadedPresentationsPubMsgHdlr.scala
@@ -20,7 +20,7 @@ trait PreuploadedPresentationsPubMsgHdlr {
       val pages = new collection.mutable.HashMap[String, PageVO]()
 
       pres.pages.foreach { p =>
-        val page = new PageVO(p.id, p.num, p.thumbUri, p.swfUri, p.txtUri, p.svgUri, p.current, p.xOffset, p.yOffset,
+        val page = new PageVO(p.id, p.num, p.thumbUri, p.txtUri, p.svgUri, p.current, p.xOffset, p.yOffset,
           p.widthRatio, p.heightRatio)
         pages += page.id -> page
       }
