+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/presentationpod/PresentationPodsApp.scala
@@ -47,7 +47,6 @@ object PresentationPodsApp {
           id = page.id,
           num = page.num,
           thumbUri = page.urls.getOrElse("thumb", ""),
-          swfUri = page.urls.getOrElse("swf", ""),
           txtUri = page.urls.getOrElse("text", ""),
           svgUri = page.urls.getOrElse("svg", ""),
           current = page.current,
@@ -80,7 +79,6 @@ object PresentationPodsApp {
         id = page.id,
         num = page.num,
         thumbUri = page.urls.getOrElse("thumb", ""),
-        swfUri = page.urls.getOrElse("swf", ""),
         txtUri = page.urls.getOrElse("text", ""),
         svgUri = page.urls.getOrElse("svg", ""),
         current = page.current,
