+++ bbb-common-message/src/main/scala/org/bigbluebutton/common2/domain/Presentation.scala
@@ -3,7 +3,7 @@ package org.bigbluebutton.common2.domain
 case class PresentationVO(id: String, temporaryPresentationId: String, name: String, current: Boolean = false,
                           pages: Vector[PageVO], downloadable: Boolean, removable: Boolean)
 
-case class PageVO(id: String, num: Int, thumbUri: String = "", swfUri: String,
+case class PageVO(id: String, num: Int, thumbUri: String = "",
                   txtUri: String, svgUri: String, current: Boolean = false, xOffset: Double = 0,
                   yOffset: Double = 0, widthRatio: Double = 100D, heightRatio: Double = 100D)
