+++ bbb-common-web/src/main/java/org/bigbluebutton/api2/IBbbWebApiGWApp.java
@@ -1,7 +1,6 @@
 package org.bigbluebutton.api2;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 import org.bigbluebutton.api.domain.BreakoutRoomsParams;
