+++ bbb-common-web/src/main/java/org/bigbluebutton/presentation/ConversionUpdateMessage.java
@@ -94,7 +94,6 @@ public class ConversionUpdateMessage {
 				Map<String, String> page = new HashMap<String, String>();
 				page.put("num", Integer.toString(i));
 				page.put("thumb", basePresUrl + "/thumbnail/" + i);
-				page.put("swf", basePresUrl + "/slide/" + i);
 				page.put("text", basePresUrl + "/textfiles/" + i);
 				
 				pages.add(page);
