+++ bbb-common-web/src/main/java/org/bigbluebutton/presentation/DocumentConversionServiceImp.java
@@ -33,7 +33,7 @@ public class DocumentConversionServiceImp implements DocumentConversionService {
 
   private IBbbWebApiGWApp gw;
   private OfficeToPdfConversionService officeToPdfConversionService;
-  private SwfSlidesGenerationProgressNotifier notifier;
+  private SlidesGenerationProgressNotifier notifier;
 
   private PresentationFileProcessor presentationFileProcessor;
 
@@ -166,7 +166,7 @@ public class DocumentConversionServiceImp implements DocumentConversionService {
     officeToPdfConversionService = s;
   }
 
-  public void setSwfSlidesGenerationProgressNotifier(SwfSlidesGenerationProgressNotifier notifier) {
+  public void setSlidesGenerationProgressNotifier(SlidesGenerationProgressNotifier notifier) {
       this.notifier = notifier;
   }
