+++ bbb-common-web/src/main/java/org/bigbluebutton/presentation/imp/PageToConvert.java
@@ -10,51 +10,39 @@ public class PageToConvert {
   private UploadedPresentation pres;
   private int page;
 
-  private boolean swfSlidesRequired;
-  private boolean svgImagesRequired;
+  private boolean svgImagesRequired=true;
   private boolean generatePngs;
   private PageExtractor pageExtractor;
 
-  private String BLANK_SLIDE;
-  private int MAX_SWF_FILE_SIZE;
 
   private TextFileCreator textFileCreator;
   private SvgImageCreator svgImageCreator;
   private ThumbnailCreator thumbnailCreator;
   private PngCreator pngCreator;
-  private PageConverter pdfToSwfConverter;
-  private SwfSlidesGenerationProgressNotifier notifier;
+  private SlidesGenerationProgressNotifier notifier;
   private File pageFile;
   private String messageErrorInConversion;
 
   public PageToConvert(UploadedPresentation pres,
                        int page,
                        File pageFile,
-                       boolean swfSlidesRequired,
                        boolean svgImagesRequired,
                        boolean generatePngs,
                        TextFileCreator textFileCreator,
                        SvgImageCreator svgImageCreator,
                        ThumbnailCreator thumbnailCreator,
                        PngCreator pngCreator,
-                       PageConverter pdfToSwfConverter,
-                       SwfSlidesGenerationProgressNotifier notifier,
-                       String blankSlide,
-                       int maxSwfFileSize) {
+                       SlidesGenerationProgressNotifier notifier) {
     this.pres = pres;
     this.page = page;
     this.pageFile = pageFile;
-    this.swfSlidesRequired = swfSlidesRequired;
     this.svgImagesRequired = svgImagesRequired;
     this.generatePngs = generatePngs;
     this.textFileCreator = textFileCreator;
     this.svgImageCreator = svgImageCreator;
     this.thumbnailCreator = thumbnailCreator;
     this.pngCreator = pngCreator;
-    this.pdfToSwfConverter = pdfToSwfConverter;
     this.notifier = notifier;
-    this.BLANK_SLIDE = blankSlide;
-    this.MAX_SWF_FILE_SIZE = maxSwfFileSize;
   }
 
   public File getPageFile() {
@@ -82,11 +70,6 @@ public class PageToConvert {
   }
 
   public PageToConvert convert() {
-
-    // Only create SWF files if the configuration requires it
-    if (swfSlidesRequired) {
-      convertPdfToSwf(pres, page, pageFile);
-    }
 
     /* adding accessibility */
     createThumbnails(pres, page, pageFile);
@@ -129,25 +112,4 @@ public class PageToConvert {
     pngCreator.createPng(pres, page, pageFile);
   }
 
-  private void convertPdfToSwf(UploadedPresentation pres, int page, File pageFile) {
-    PdfToSwfSlide slide = setupSlide(pres, page, pageFile);
-    generateSlides(pres, slide);
-  }
-
-
-  private void generateSlides(UploadedPresentation pres, PdfToSwfSlide slide) {
-    slide.createSlide();
-    if (!slide.isDone()) {
-      slide.generateBlankSlide();
-    }
-  }
-
-  private PdfToSwfSlide setupSlide(UploadedPresentation pres, int page, File pageFile) {
-    PdfToSwfSlide slide = new PdfToSwfSlide(pres, page, pageFile);
-    slide.setBlankSlide(BLANK_SLIDE);
-    slide.setMaxSwfFileSize(MAX_SWF_FILE_SIZE);
-    slide.setPageConverter(pdfToSwfConverter);
-
-    return slide;
-  }
 }
