+++ bbb-common-web/src/main/java/org/bigbluebutton/presentation/imp/SvgImageCreatorImp.java
@@ -28,7 +28,7 @@ import com.zaxxer.nuprocess.NuProcessBuilder;
 public class SvgImageCreatorImp implements SvgImageCreator {
     private static Logger log = LoggerFactory.getLogger(SvgImageCreatorImp.class);
 
-    private SwfSlidesGenerationProgressNotifier notifier;
+    private SlidesGenerationProgressNotifier notifier;
     private long imageTagThreshold;
     private long pathsThreshold;
     private int convPdfToSvgTimeout = 60;
@@ -381,8 +381,8 @@ public class SvgImageCreatorImp implements SvgImageCreator {
         pathsThreshold = threshold;
     }
     
-    public void setSwfSlidesGenerationProgressNotifier(
-        SwfSlidesGenerationProgressNotifier notifier) {
+    public void setSlidesGenerationProgressNotifier(
+        SlidesGenerationProgressNotifier notifier) {
       this.notifier = notifier;
     }
