+++ bbb-common-web/src/main/scala/org/bigbluebutton/api2/MsgBuilder.scala
@@ -68,13 +68,12 @@ object MsgBuilder {
     val id = presId + "/" + page
     val current = if (page == 1) true else false
     val thumbUrl = presBaseUrl + "/thumbnail/" + page
-    val swfUrl = presBaseUrl + "/slide/" + page
 
     val txtUrl = presBaseUrl + "/textfiles/" + page
     val svgUrl = presBaseUrl + "/svg/" + page
     val pngUrl = presBaseUrl + "/png/" + page
 
-    val urls = Map("swf" -> swfUrl, "thumb" -> thumbUrl, "text" -> txtUrl, "svg" -> svgUrl, "png" -> pngUrl)
+    val urls = Map("thumb" -> thumbUrl, "text" -> txtUrl, "svg" -> svgUrl, "png" -> pngUrl)
 
     PresentationPageConvertedVO(
       id = id,
@@ -164,14 +163,12 @@ object MsgBuilder {
       val num = i
       val current = if (i == 1) true else false
       val thumbnail = presBaseUrl + "/thumbnail/" + i
-      val swfUri = presBaseUrl + "/slide/" + i
 
       val txtUri = presBaseUrl + "/textfiles/" + i
       val svgUri = presBaseUrl + "/svg/" + i
 
-      val p = PageVO(id = id, num = num, thumbUri = thumbnail, swfUri = swfUri,
-        txtUri = txtUri, svgUri = svgUri,
-        current = current)
+      val p = PageVO(id = id, num = num, thumbUri = thumbnail,
+        txtUri = txtUri, svgUri = svgUri, current = current)
       pages += p.id -> p
     }
