+++ bigbluebutton-web/grails-app/conf/bigbluebutton.properties
@@ -27,17 +27,8 @@ appLogLevel=DEBUG
 presentationDir=/var/bigbluebutton
 
 #----------------------------------------------------
-# Directory where SWFTOOLS (pdf2swf, jpeg2swf, png2swf) are located
-swfToolsDir=/usr/bin
-
-#----------------------------------------------------
 # Directory where ImageMagick's convert executable is located
 imageMagickDir=/usr/bin
-
-#----------------------------------------------------
-# Fonts directory passed into PDF2SWF to support highlighting of texts
-# in the SWF slides.
-fontsDir=/usr/share/fonts
 
 #----------------------------------------------------
 # Executable for presentation checker
@@ -55,7 +46,6 @@ presOfficeConversionExec=/usr/share/bbb-libreoffice-conversion/convert.sh
 #----------------------------------------------------
 # These will be copied in cases where the conversion process
 # fails to generate a slide from the uploaded presentation
-BLANK_SLIDE=/usr/share/bigbluebutton/blank/blank-slide.swf
 BLANK_PRESENTATION=/usr/share/bigbluebutton/blank/blank-presentation.pdf
 BLANK_THUMBNAIL=/usr/share/bigbluebutton/blank/blank-thumb.png
 BLANK_PNG=/usr/share/bigbluebutton/blank/blank-png.png
@@ -75,18 +65,11 @@ maxNumPages=200
 maxFileSizeUpload=30000000
 
 #----------------------------------------------------
-# Maximum swf file size for load to the client (default 500000).
-MAX_SWF_FILE_SIZE=500000
-
-#----------------------------------------------------
-# Maximum allowed number of place object tags in the converted SWF, if exceeded the conversion will fallback to full BMP (default 800)
+# Maximum allowed number of place object tags in generated svg, if exceeded the conversion will fallback to full BMP (default 800)
 placementsThreshold=800
 
-# Maximum allowed number of bitmap images in the converted SWF, if exceeded the conversion will fallback to full BMP (default 800)
+# Maximum allowed number of bitmap images in generated svg, if exceeded the conversion will fallback to full BMP (default 800)
 imageTagThreshold=800
-
-# Maximum allowed number of define text tags in the converted SWF, if exceeded the conversion will fallback to full BMP (default 200)
-defineTextThreshold=2000
 
 #------------------------------------
 # Number of threads in the pool to do the presentation conversion.
@@ -145,16 +128,6 @@ officeToPdfConversionTimeout=60
 officeToPdfMaxConcurrentConversions=4
 
 #----------------------------------------------------
-# Conversion of the presentation slides to SWF to be
-# used in the Flash client
-swfSlidesRequired=false
-
-#----------------------------------------------------
-# Additional conversion of the presentation slides to SVG
-# to be used in the HTML5 client
-svgImagesRequired=true
-
-#----------------------------------------------------
 # Additional conversion of the presentation slides to PNG
 # to be used in the IOS mobile client
 generatePngs=false
@@ -164,8 +137,9 @@ pngSlideWidth=1200
 defaultNumDigitsForTelVoice=5
 
 #----------------------------------------------------
-# Configuration for large images, 2 MB by default, if bigger it will down-scaled
-maxImageSize=2000000
+# Maximum image resolution, if image exceeds this limit it will be resized
+maxImageWidth=2048
+maxImageHeight=1536
 
 #----------------------------------------------------
 # Configuration for large PDF, 14 MB by default, if bigger it will be analysed during the conversion process
